Complete homework 14_1?
