+++ README.md
@@ -1,0 +1,2 @@
+# Домашняя работа Блока ООП
+## Цель проекта
